requirements.txt
crawl4ai>=0.4.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
playwright>=1.40.0
supabase>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
email-validator>=2.0.0
openai>=1.0.0
